RDP-2 add all face card component tags
RDP-2 icon files added

RDP-2 add missing index.html file

RDM-2 add icons and organize files

RDP-2 remove unused files
