+++ src/components/facecard/faceCard.css
@@ -1,0 +1,67 @@
+.card-body {
+  width: 50.5vw;
+  height: 62vh;
+  position: absolute;
+  background: linear-gradient(to bottom, #F1F5F8 0%, #F1F5F8 29%, #B4B8BA 29%, #B4B8BA 29.2%, #FFFFFF 29.2%, #FFFFFF 100%);
+  border-radius: 4px;
+  margin-top: 6vh;
+  box-shadow: 0 2px 6px 0 rgba(0, 0, 0, 0.2), 0 4px 10px 0 rgba(0, 0, 0, 0.19);
+  align-items: center;
+  justify-content: center;
+  display: flex;
+  flex-direction: column;
+}
+
+.image-div {
+  height: 140px;
+  width: 140px;
+  background-color: white;
+  border-radius: 50%;
+  margin-bottom: 8px;
+  box-shadow: 0 2px 6px 0 rgba(0, 0, 0, 0.2), 0 4px 10px 0 rgba(0, 0, 0, 0.19);
+}
+
+.user-info {
+  margin-top: 15px;
+  font-size: large;
+  max-height: 100px;
+}
+
+.user-info h1 {
+  margin-top: -5px;
+}
+
+.icons-container {
+  width: 47vw;
+  height: 10vh;
+  display: grid;
+  grid-template-columns: repeat(6, 16.6%);
+  align-items: center;
+  justify-items: center;
+}
+
+.icons {
+  width: 2.5vw;
+  height: 4vh;
+
+}
+
+.random-button {
+  background-color: #48A6E8;
+  color: white;
+  height: 35px;
+  width: 150px;
+  border-radius: 3px;
+  text-transform: uppercase;
+  font-size: 15px;
+  margin: 10px;
+  margin-bottom: 3px;
+  border: none;
+  cursor: pointer;
+  transition: background 2s ease;
+}
+
+.random-button:hover {
+  color: #49a6e9;
+  background-color: #063251;
+}
