--- src/facCard.js
@@ -1,26 +1,0 @@
-import React from "react";
-import './faceCard.css';
-
-const FaceCard = () => {
-  return (
-    <div className="card-body">
-      <div className="image-div">
-      
-      </div>
-
-      <div className="user-info">
-
-      </div>
-
-      <div className="icons">
-
-      </div>
-
-      <div className="random-button">
-
-      </div>
-    </div>
-  );
-};
-
-export default FaceCard;
